+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.4] - TODO:修改了onTap事件
+* TODO: CupertionPopoverButton添加了onTap事件,返回True不打开Popover
+* TODO: CupertinoPopoverMenuItem修改了onTap事件,返回True不关闭Popover
+
 ## [0.1.1] - TODO:添加了WeuiToast效果,完善了文档信息
 * TODO: 添加了WeuiToast效果
 * TODO: 完善了文档信息